requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
distro==1.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
openai==1.2.4
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
setuptools==68.2.2
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
wheel==0.41.3
